Promote dialog overlays to CircuitX
